requirements.txt
numpy
pandas
redis
opencv-python-headless
scikit-learn
streamlit
streamlit-webrtc
insightface
onnxruntime
streamlit-authenticator
flake8
black
pymongo
pydantic-settings
python-dotenv